+++ edition/voxel_tool_terrain.h
@@ -53,19 +53,6 @@ public:
 
 	void run_blocky_random_tick(AABB voxel_area, int voxel_count, const Callable &callback, int block_batch_count);
 
-	// For easier unit testing (the regular one needs a terrain setup etc, harder to test atm)
-	// The `_static` suffix is because it otherwise conflicts with the non-static method when registering the class
-	static void run_blocky_random_tick_static(
-			VoxelData &data,
-			Box3i voxel_box,
-			const VoxelBlockyLibraryBase &lib,
-			RandomPCG &random,
-			int voxel_count,
-			int batch_count,
-			void *callback_data,
-			bool (*callback)(void *, Vector3i, int64_t)
-	);
-
 	void for_each_voxel_metadata_in_area(AABB voxel_area, const Callable &callback);
 
 protected: